[REFACTOR] FamilyProfile삭제 후 home-page로 이동
